Move the loading function onto timmers
Move the functions that pull the data for each view on to timers so
they will update the view if new data is pulled from the server or if
an item becomes live. Also move the function that pulls the data from
the server into a background task so it does not hang the app while it
waits for the server to respond.